Removed robot, added Test
