+++ settings.gradle
@@ -38,6 +38,6 @@ include 'WML-Rev'
 include 'UDP_TransferNT'
 
 // Robot includes (not published)
-include "Robot"
+include "Test"
 
 rootProject.name = 'WML'
